+++ package.js
@@ -1,15 +1,14 @@
 Package.describe({
-  name: "emmerge:google-fake",
+  name: "ryepdx:linkedin-fake",
   summary: "A fake for use in testing. Fakes the oauth calls amongst other APIs.",
-  version: "0.0.4_1",
-  git: "https://github.com/xolvio/meteor-google-fake",
+  version: "0.0.4",
+  git: "https://github.com/ryepdx/meteor-linkedin-fake",
   debugOnly: true
 });
 
 Package.on_use(function (api) {
-  api.use('google@1.1.2', ['client', 'server']);
   api.use(['xolvio:http-interceptor@0.4.0'], ['server']);
   api.use(['iron:router@1.0.6'], ['server']);
   api.add_files('oauth-fake-client.js', 'client');
-  api.add_files('google-fake.js', ['server']);
+  api.add_files('linkedin-fake.js', ['server']);
 });